+++ Setup/UpgradeSchema.php
@@ -282,29 +282,6 @@ class UpgradeSchema implements UpgradeSchemaInterface
                 );
         }
 
-        // create attribute parameter to include/exclude from product feed
-        $catalogEavAttributeTable = $installer->getTable('catalog_eav_attribute');
-        if (
-            $installer->tableExists($catalogEavAttributeTable)
-            && !$installer->getConnection()->tableColumnExists(
-                $catalogEavAttributeTable, 'include_in_unbxd_product_feed'
-            )
-        ) {
-            $installer->getConnection()
-                ->addColumn(
-                    $catalogEavAttributeTable,
-                    'include_in_unbxd_product_feed',
-                    [
-                        'type' => Table::TYPE_SMALLINT,
-                        'nullable' => false,
-                        'unsigned' => true,
-                        'length' => 1,
-                        'default' => 1,
-                        'comment' => 'Include In Unbxd Product Feed'
-                    ]
-                );
-        }
-
         $installer->endSetup();
     }
 }